Function and type index

Parameters and constants

SpeedyWeather.Parameters — Type

P = Parameters{M<:ModelSetup}(kwargs...) <: AbstractParameters{M}

A struct to hold all model parameters that may be changed by the user. The struct uses keywords such that default values can be changed at creation. The default values of the keywords define the default model setup.

NF::DataType: number format

trunc::Int64: spectral truncation

Grid::Type{<:SpeedyWeather.RingGrids.AbstractGrid}: grid in use

dealiasing::Float64: dealiasing factor, 1=linear, 2=quadratic, 3=cubic grid

planet::SpeedyWeather.Planet: planet

mol_mass_dry_air::Any: molar mass of dry air [g/mol]

mol_mass_vapour::Any: molar mass of water vapour [g/mol]

cₚ::Float64: specific heat at constant pressure [J/K/kg]

R_gas::Float64: universal gas constant [J/K/mol]

R_dry::Float64: specific gas constant for dry air [J/kg/K]

R_vapour::Float64: specific gas constant for water vapour [J/kg/K]

alhc::Float64: latent heat of condensation [J/g] for consistency with specific humidity [g/Kg]

alhs::Float64: latent heat of sublimation [?]

sbc::Float64: stefan-Boltzmann constant [W/m²/K⁴]

lapse_rate::Float64: moist adiabatic temperature lapse rate $-dT/dz$ [K/km]

temp_ref::Float64: absolute temperature at surface $z=0$ [K]

temp_top::Float64: absolute temperature in stratosphere [K]

ΔT_stratosphere::Float64: for stratospheric lapse rate [K] after Jablonowski

scale_height::Float64: scale height for pressure [km]

pres_ref::Float64: surface pressure [hPa]

scale_height_humid::Float64: scale height for specific humidity [km]

relhumid_ref::Float64: relative humidity of near-surface air [1]

water_pres_ref::Float64: saturation water vapour pressure [Pa]

layer_thickness::Float64: layer thickness for the shallow water model [km]

GLcoefs::SpeedyWeather.Coefficients: vertical coordinates of the nlev vertical levels, defined by a generalised logistic function, interpolating ECMWF's L31 configuration

σ_tropopause::Float64: σ coordinate where the tropopause starts

σ_levels_half::Vector{Float64}: only used if set manually, otherwise empty

nlev::Int64: number of vertical levels

diffusion::SpeedyWeather.DiffusionParameters: horizontal (hyper)-diffusion

vertical_diffusion::SpeedyWeather.VerticalDiffusion: vertical diffusion

static_energy_diffusion::SpeedyWeather.VerticalDiffusion: static energy diffusion

interface_relaxation::Bool: turn on interface relaxation for shallow water?

interface_relax_time::Float64: time scale [hrs] of interface relaxation

interface_relax_amplitude::Float64: Amplitude [m] of interface relaxation

physics::Bool: en/disables the physics parameterizations

n_shortwave::Int64: Compute shortwave radiation every n steps

sppt_on::Bool: Turn on SPPT?

magnus_coefs::SpeedyWeather.Coefficients: For computing saturation vapour pressure

k_lsc::Int64: Index of atmospheric level at which large-scale condensation begins

RH_thresh_pbl_lsc::Float64: Relative humidity threshold for boundary layer

RH_thresh_range_lsc::Float64: Vertical range of relative humidity threshold

RH_thresh_max_lsc::Float64: Maximum relative humidity threshold

humid_relax_time_lsc::Float64: Relaxation time for humidity (hours)

pres_thresh_cnv::Float64: Minimum (normalised) surface pressure for the occurrence of convection

RH_thresh_pbl_cnv::Float64: Relative humidity threshold for convection in PBL

RH_thresh_trop_cnv::Float64: Relative humidity threshold for convection in the troposphere

humid_relax_time_cnv::Float64: Relaxation time for PBL humidity (hours)

max_entrainment::Float64: Maximum entrainment as a fraction of cloud-base mass flux

ratio_secondary_mass_flux::Float64: Ratio between secondary and primary mass flux at cloud-base

nband::Int64: Number of bands used to compute fband

radiation_coefs::SpeedyWeather.Coefficients: radiation coefficients

boundary_layer::SpeedyWeather.BoundaryLayer{Float64}: boundary layer drag

temperature_relaxation::SpeedyWeather.TemperatureRelaxation{Float64}: temperature relaxation

startdate::Dates.DateTime: time at which the integration starts

n_days::Float64: number of days to integrate for

Δt_at_T31::Float64: time step in minutes for T31, scale linearly to trunc

robert_filter::Float64: Robert (1966) time filter coefficeint to suppress comput. mode

williams_filter::Float64: William's time filter (Amezcua 2011) coefficient for 3rd order acc

implicit_α::Float64: coefficient for semi-implicit computations to filter gravity waves

recalculate_implicit::Int64: recalculate implicit operators on temperature profile every n time steps

recompute_legendre::Bool: recomputation is slower but requires less memory

legendre_NF::DataType: which format to use to calculate the Legendre polynomials

legendre_shortcut::Symbol: :linear, :quadratic, :cubic, :lincub_coslat, :linquad_coslat²

boundary_path::String: package location is default

orography::SpeedyWeather.AbstractOrography: orography

orography_scale::Float64: scale orography by a factor

orography_path::String: path of orography

orography_file::String: filename of orography

seed::Int64: random seed for the global random number generator

initial_conditions::SpeedyWeather.InitialConditions: initial conditions

pressure_on_orography::Bool: calculate the initial surface pressure from orography

verbose::Bool: print dialog for feedback

debug::Bool: print debug info, NaR detection

output::Bool: Store data in netCDF?

output_dt::Float64: output time step [hours]

output_path::String: path to output folder

run_id::Union{Int64, String}: name of the output folder, defaults to 4-digit number counting up from run-0001

output_filename::String: name of the output netcdf file

output_vars::Vector{Symbol}: variables to output: :u, :v, :vor, :div, :temp, :humid

compression_level::Int64: compression level; 1=low but fast, 9=high but slow

keepbits::NamedTuple: mantissa bits to keep for every variable

version::VersionNumber: SpeedyWeather.jl version number

output_NF::DataType: number format used for output

output_nlat_half::Int64: 0 = reuse nlat_half from dynamical core

output_Grid::Type{<:SpeedyWeather.RingGrids.AbstractFullGrid}: output grid

output_Interpolator::Type{<:SpeedyWeather.RingGrids.AbstractInterpolator}: output interpolator

output_matrix::Bool: if true sort gridpoints into a matrix

output_quadrant_rotation::NTuple{4, Int64}: rotation of output quadrant

output_matrix_quadrant::NTuple{4, Tuple{Int64, Int64}}: matrix of output quadrant

missing_value::Float64: missing value to be used in netcdf output

write_restart::Bool: also write restart file if output==true?

restart_path::String: path for restart file

restart_id::Union{Int64, String}: run_id of restart file in run-????/restart.jld2

Missing docstring.

Missing docstring for SpeedyWeather.Constants. Check Documenter's build log for details.

Boundaries and boundary conditions

Missing docstring.

Missing docstring for SpeedyWeather.Boundaries. Check Documenter's build log for details.

Spherical harmonic transform

Missing docstring.

Missing docstring for SpeedyWeather.GeoSpectral. Check Documenter's build log for details.

SpeedyWeather.SpeedyTransforms.SpectralTransform — Type

S = SpectralTransform{NF<:AbstractFloat}(...)

SpectralTransform struct that contains all parameters and preallocated arrays for the spectral transform.

SpeedyWeather.SpeedyTransforms.spectral — Function

alms = spectral(    map::AbstractMatrix;
                    Grid::Type{<:AbstractGrid}=DEFAULT_GRID,
                    kwargs...)

Converts map to grid(map) to execute spectral(map::AbstractGrid;kwargs...).

alms = spectral(    map::AbstractGrid;
                    Grid::Type{<:AbstractGrid}=DEFAULT_GRID,
                    kwargs...)

Converts map to Grid(map) to execute spectral(map::AbstractGrid;kwargs...).

alms = spectral(    map::AbstractMatrix;
                    Grid::Type{<:AbstractGrid}=DEFAULT_GRID,
                    kwargs...)

Spectral transform (grid to spectral) map to grid(map) to execute spectral(map::AbstractGrid;kwargs...).

SpeedyWeather.SpeedyTransforms.spectral! — Function

spectral!(  alms::LowerTriangularMatrix,
            map::AbstractGrid,
            S::SpectralTransform)

Spectral transform (grid to spectral space) from the gridded field map on a grid<:AbstractGrid to a LowerTriangularMatrix of spherical harmonic coefficients alms. Uses FFT in the zonal direction, and a Legendre Transform in the meridional direction exploiting symmetries. The spectral transform is number format-flexible as long as the parametric types of map, alms, S are identical. The spectral transform is grid-flexible as long as the typeof(map)<:AbstractGrid. Uses the precalculated arrays, FFT plans and other constants in the SpectralTransform struct S.

SpeedyWeather.SpeedyTransforms.gridded — Function

map = gridded(  alms::AbstractMatrix;
                recompute_legendre::Bool=true,
                grid::Type{<:AbstractGrid}=DEFAULT_GRID)

Spectral transform (spectral to grid space) from spherical coefficients alms to a newly allocated gridded field map. Based on the size of alms the grid type grid, the spatial resolution is retrieved based on the truncation defined for grid. SpectralTransform struct S is allocated to execute gridded(alms,S).

map = gridded(  alms::AbstractMatrix,
                S::SpectralTransform)

Spectral transform (spectral to grid space) from spherical coefficients alms to a newly allocated gridded field map with precalculated properties based on the SpectralTransform struct S. alms is converted to a LowerTriangularMatrix to execute the in-place gridded!.

SpeedyWeather.SpeedyTransforms.gridded! — Function

gridded!(   diagn::DiagnosticVariables{NF}, # all diagnostic variables
            progn::PrognosticVariables{NF}, # all prognostic variables
            M::BarotropicModel,             # everything that's constant
            lf::Int=1                       # leapfrog index
            ) where NF

Propagate the spectral state of the prognostic variables progn to the diagnostic variables in diagn for the barotropic vorticity model. Updates grid vorticity, spectral stream function and spectral and grid velocities u,v.

gridded!(   diagn::DiagnosticVariables{NF}, # all diagnostic variables
            progn::PrognosticVariables{NF}, # all prognostic variables
            lf::Int=1                       # leapfrog index
            M::ShallowWaterModel,           # everything that's constant
            ) where NF

Propagate the spectral state of the prognostic variables progn to the diagnostic variables in diagn for the shallow water model. Updates grid vorticity, grid divergence, grid interface displacement (pres_grid) and the velocities U,V (scaled by cos(lat)).

gridded!(   map::AbstractGrid,
            alms::LowerTriangularMatrix,
            S::SpectralTransform)

Spectral transform (spectral to grid) of the spherical harmonic coefficients alms to a gridded field map. The spectral transform is number format-flexible as long as the parametric types of map, alms, S are identical. The spectral transform is grid-flexible as long as the typeof(map)<:AbstractGrid. Uses the precalculated arrays, FFT plans and other constants in the SpectralTransform struct S.

Missing docstring.

Missing docstring for SpeedyWeather.triangular_truncation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.roundup_fft. Check Documenter's build log for details.

SpeedyWeather.SpeedyTransforms.spectral_truncation — Function

alms_trunc = spectral_truncation(alms,trunc)

Returns a spectral coefficient matrix alms_trunc that is truncated from alms to the size (trunc+1)². alms_trunc only contains those coefficient of alms for which m,l ≤ trunc, and l ≥ m are zero anyway. If trunc is larger than the implicit truncation in alms obtained from its size than spectral_interpolation is automatically called instead, returning alms_interp, a coefficient matrix that is larger than alms with padded zero coefficients.

SpeedyWeather.SpeedyTransforms.spectral_truncation! — Function

spectral_truncation!(alms::AbstractMatrix,ltrunc::Integer,mtrunc::Integer)

Truncate spectral coefficients alms in-place by setting (a) the upper right triangle to zero and (b) all coefficients for which the degree l is larger than the truncation ltrunc or order m larger than the truncaction mtrunc.

spectral_truncation!(alms::LowerTriangularMatrix,ltrunc::Integer,mtrunc::Integer)

Truncate spectral coefficients alms in-place by setting all coefficients for which the degree l is larger than the truncation ltrunc or order m larger than the truncaction mtrunc. Similar to spectral_truncation!(::AbstractMatrix, ...) but skips the upper triangle which is zero by design for LowerTriangularMatrix.

spectral_truncation!(alms,trunc)

Truncate spectral coefficients alms in-place by setting (a) the upper right triangle to zero and (b) all coefficients for which the degree l is larger than the truncation trunc.

spectral_truncation!(alms)

Truncate spectral coefficients alms in-place by setting the upper right triangle to zero. This is to enforce that all coefficients for which the degree l is larger than order m are zero.

Missing docstring.

Missing docstring for SpeedyWeather.spectral_interpolation!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.get_legendre_polynomials!. Check Documenter's build log for details.

SpeedyWeather.SpeedyTransforms.∇²! — Function

∇²!(    ∇²alms::LowerTriangularMatrix,
        alms::LowerTriangularMatrix,
        S::SpectralTransform;
        add::Bool=false,
        flipsign::Bool=false,
        inverse::Bool=false)

Laplace operator ∇² applied to the spectral coefficients alms in spherical coordinates. The radius R is omitted in the eigenvalues which are precomputed in S. ∇²! is the in-place version which directly stores the output in the first argument ∇²alms.

Keyword arguments

add=true adds the ∇²(alms) to the output

flipsign=true computes -∇²(alms) instead

inverse=true computes ∇⁻²(alms) instead

Default is add=false, flipsign=false, inverse=false. These options can be combined.

Missing docstring.

Missing docstring for SpeedyWeather.∇². Check Documenter's build log for details.

SpeedyWeather.SpeedyTransforms.∇⁻²! — Function

∇⁻²!(   ∇⁻²alms::LowerTriangularMatrix,
        alms::LowerTriangularMatrix,
        S::SpectralTransform;
        add::Bool=false,
        flipsign::Bool=false)

Calls ∇²!(∇⁻²alms, alms, S; add, flipsign, inverse=true).

Missing docstring.

Missing docstring for SpeedyWeather.∇⁻². Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.gradient_latitude!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.gradient_latitude. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.gradient_longitude!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.gradient_longitude. Check Documenter's build log for details.

SpeedyWeather.SpeedyTransforms.divergence! — Function

divergence!(div::LowerTriangularMatrix,
            u::LowerTriangularMatrix,
            v::LowerTriangularMatrix,
            S::SpectralTransform{NF};
            flipsign::Bool=false,
            add::Bool=false,
            )

Divergence of a vector u,v written into div, div = ∇⋅(u,v). u,v are expected to have a 1/coslat-scaling included, then div is not scaled. flipsign option calculates -∇⋅(u,v) instead. add option calculates div += ∇⋅(u,v) instead. flipsign and add can be combined. This functions only creates the kernel and calls the generic divergence function _divergence! subsequently.

SpeedyWeather.SpeedyTransforms.curl! — Function

curl!(  curl::LowerTriangularMatrix,
        u::LowerTriangularMatrix,
        v::LowerTriangularMatrix,
        S::SpectralTransform;
        flipsign::Bool=false,
        add::Bool=false,
        )

Curl of a vector u,v written into curl, curl = ∇×(u,v). u,v are expected to have a 1/coslat-scaling included, then curl is not scaled. flipsign option calculates -∇×(u,v) instead. add option calculates curl += ∇×(u,v) instead. flipsign and add can be combined. This functions only creates the kernel and calls the generic divergence function _divergence! subsequently with flipped u,v -> v,u for the curl.

Missing docstring.

Missing docstring for SpeedyWeather._divergence!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.curl_div!. Check Documenter's build log for details.

SpeedyWeather.SpeedyTransforms.UV_from_vordiv! — Function

UV_from_vordiv!(U::LowerTriangularMatrix,
                V::LowerTriangularMatrix,
                vor::LowerTriangularMatrix,
                div::LowerTriangularMatrix,
                S::SpectralTransform)

Get U,V (=(u,v)*coslat) from vorticity ζ and divergence D in spectral space. Two operations are combined into a single linear operation. First, invert the spherical Laplace ∇² operator to get stream function from vorticity and velocity potential from divergence. Then compute zonal and meridional gradients to get U,V.

SpeedyWeather.SpeedyTransforms.UV_from_vor! — Function

UV_from_vor!(   U::LowerTriangularMatrix,
                V::LowerTriangularMatrix,
                vor::LowerTriangularMatrix,
                S::SpectralTransform)

Get U,V (=(u,v)*coslat) from vorticity ζ spectral space (divergence D=0) Two operations are combined into a single linear operation. First, invert the spherical Laplace ∇² operator to get stream function from vorticity. Then compute zonal and meridional gradients to get U,V.

Missing docstring.

Missing docstring for SpeedyWeather.ϵlm. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.get_recursion_factors. Check Documenter's build log for details.

Dynamics

SpeedyWeather.bernoulli_potential! — Function

bernoulli_potential!(   diagn::DiagnosticVariablesLayer, 
                        G::Geometry,
                        S::SpectralTransform)

Computes the Laplace operator ∇² of the Bernoulli potential B in spectral space.

computes the kinetic energy KE = ½(u²+v²) on the grid

transforms KE to spectral space

adds geopotential for the Bernoulli potential in spectral space

takes the Laplace operator.

This version is used for both ShallowWater and PrimitiveEquation, only the geopotential calculation in geopotential! differs.

SpeedyWeather.volume_flux_divergence! — Function

volume_flux_divergence!(diagn::DiagnosticVariablesLayer,
                        surface::SurfaceVariables,
                        model::ShallowWater)

Computes the (negative) divergence of the volume fluxes uh,vh for the continuity equation, -∇⋅(uh,vh).

Missing docstring.

Missing docstring for SpeedyWeather.vorticity_fluxes!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.vorticity_flux_curl!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.vorticity_flux_divergence!. Check Documenter's build log for details.

Geometry

SpeedyWeather.Geometry — Type

Geometry{NF<:AbstractFloat} <: AbstractGeometry

Geometry struct containing parameters and arrays describing an iso-latitude grid <:AbstractGrid and the vertical levels. NF is the number format used for the precomputed constants.

SpeedyWeather.vertical_coordinates — Function

σ_levels_half = vertical_coordinates(P::Parameters)

Vertical sigma coordinates defined by their nlev+1 half levels σ_levels_half. Sigma coordinates are fraction of surface pressure (p/p0) and are sorted from top (stratosphere) to bottom (surface). The first half level is at 0 the last at 1. Evaluate a generalised logistic function with coefficients in P for the distribution of values in between. Default coefficients follow the L31 configuration historically used at ECMWF.

SpeedyWeather.GenLogisticCoefs — Type

Coefficients of the generalised logistic function to describe the vertical coordinate. Default coefficients A,K,C,Q,B,M,ν are fitted to the old L31 configuration at ECMWF. See geometry.jl and function vertical_coordinate for more informaiton.

Following the notation of https://en.wikipedia.org/wiki/Generalisedlogisticfunction (Dec 15 2021).

Change default parameters for more/fewer levels in the stratosphere vs troposphere vs boundary layer.

SpeedyWeather.generalised_logistic — Function

Generalised logistic function based on the coefficients in coefs.

Time stepping

SpeedyWeather.time_stepping! — Function

time_stepping!( progn::PrognosticVariables,     # all prognostic variables
                diagn::DiagnosticVariables,     # all pre-allocated diagnostic variables
                model::ModelSetup)              # all precalculated structs

Main time loop that that initialises output and feedback, loops over all time steps and calls the output and feedback functions.

SpeedyWeather.timestep! — Function

timestep!(  progn::PrognosticVariables,     # all prognostic variables
            diagn::DiagnosticVariables,     # all pre-allocated diagnostic variables
            time::DateTime,                 # time at timestep
            dt::Real,                       # time step (mostly =2Δt, but for init steps =Δt,Δt/2)
            lf1::Int=2,                     # leapfrog index 1 (dis/enables Robert+William's filter)
            lf2::Int=2,                     # leapfrog index 2 (time step used for tendencies)
            M::BarotropicModel,             # everything that's constant at runtime
            )

Calculate a single time step for the barotropic vorticity equation model of SpeedyWeather.jl

timestep!(  progn::PrognosticVariables{NF}, # all prognostic variables
            diagn::DiagnosticVariables{NF}, # all pre-allocated diagnostic variables
            time::DateTime,                 # time at timestep
            dt::Real,                       # time step (mostly =2Δt, but for init steps =Δt,Δt/2)
            M::ShallowWaterModel,           # everything that's constant at runtime
            lf1::Int=2,                     # leapfrog index 1 (dis/enables Robert+William's filter)
            lf2::Int=2                      # leapfrog index 2 (time step used for tendencies)
            ) where {NF<:AbstractFloat}

Calculate a single time step for the shallow water model of SpeedyWeather.jl

timestep!(  progn::PrognosticVariables{NF}, # all prognostic variables
            diagn::DiagnosticVariables{NF}, # all pre-allocated diagnostic variables
            time::DateTime,                 # time at timestep
            dt::Real,                       # time step (mostly =2Δt, but for init steps =Δt,Δt/2)
            M::PrimitiveEquation,      # everything that's constant at runtime
            lf1::Int=2,                     # leapfrog index 1 (dis/enables Robert+William's filter)
            lf2::Int=2                      # leapfrog index 2 (time step used for tendencies)
            ) where {NF<:AbstractFloat}

Calculate a single time step for the primitive equation model of SpeedyWeather.jl

SpeedyWeather.first_timesteps! — Function

first_timesteps!(   progn::PrognosticVariables, # all prognostic variables
                    diagn::DiagnosticVariables, # all pre-allocated diagnostic variables
                    time::DateTime,             # time at timestep
                    M::ModelSetup,              # everything that is constant at runtime
                    feedback::AbstractFeedback  # feedback struct
                    )

Performs the first two initial time steps (Euler forward, unfiltered leapfrog) to populate the prognostic variables with two time steps (t=0,Δt) that can then be used in the normal leap frogging.

SpeedyWeather.leapfrog! — Function

leapfrog!(  A_old::LowerTriangularMatrix{Complex{NF}},      # prognostic variable at t
            A_new::LowerTriangularMatrix{Complex{NF}},      # prognostic variable at t+dt
            tendency::LowerTriangularMatrix{Complex{NF}},   # tendency (dynamics+physics) of A
            dt::Real,                                       # time step (=2Δt, but for init steps =Δt,Δt/2)
            lf::Int=2,                                      # leapfrog index to dis/enable William's filter
            C::DynamicsConstants{NF},                       # struct with constants used at runtime
            ) where {NF<:AbstractFloat}                     # number format NF

Performs one leapfrog time step with (lf=2) or without (lf=1) Robert+William's filter (see William (2009), Montly Weather Review, Eq. 7-9).

Longwave radiation

SpeedyWeather.radset! — Function

function radset!(model::PrimitiveEquation) where {NF<:AbstractFloat}

Compute energy fractions in four longwave bands as a function of temperature.

SpeedyWeather.radlw_down! — Function

function radlw_down!(
    column::ColumnVariables{NF}, model::PrimitiveEquation
) where {NF<:AbstractFloat}

Compute the downward flux of longwave radiation. Inputs variables are temp,wvi,tau2. Output column varables arefsfcd,dfabs,flux,st4a`.

SpeedyWeather.compute_bbe! — Function

function compute_bbe!(
    column::ColumnVariables{NF}, model::PrimitiveEquation
) where {NF<:AbstractFloat}

Computes black-body (or grey-body) emissions.

Input and output variables are ts and fsfcu, respectively.

SpeedyWeather.radlw_up! — Function

function radlw_up!(
    column::ColumnVariables{NF}, model::PrimitiveEquation
) where {NF<:AbstractFloat}

Computes the upward flux of longwave radiation.

Input variables are nlev, temp, fsfcu, fsfcd, flux, ts, tau2, st4a, dfabs, stratc, σ_levels_thick, n_stratosphere_levels. Output column variables are fsfc and ftop.

Shortwave radiation

SpeedyWeather.shortwave_radiation! — Function

function shortwave_radiation!(
    column::ColumnVariables{NF}, model::PrimitiveEquation
)

Compute air temperature tendencies from shortwave radiation for an atmospheric column. For more details see http://users.ictp.it/~kucharsk/speedydescription/kmver41_appendixA.pdf

SpeedyWeather.solar! — Function

function solar!(column::ColumnVariables{NF})

Compute average daily flux of solar radiation for an atmospheric column, from Hartmann (1994).

SpeedyWeather.sol_oz! — Function

function sol_oz!(
    column::ColumnVariables{NF}, model::PrimitiveEquation
)

Compute solar radiation parametres for an atmospheric column.

SpeedyWeather.cloud! — Function

function cloud!(
    column::ColumnVariables{NF}, model::PrimitiveEquation
)

Compute shortwave radiation cloud contibutions for an atmospheric column.

SpeedyWeather.radsw! — Function

function radsw!(
    column::ColumnVariables{NF}, model::PrimitiveEquation
)

Compute shortwave radiation fluxes for an atmospheric column.