How to run SpeedyWeather.jl

The simplest way to run SpeedyWeather.jl with default parameters is

using SpeedyWeather
run_speedy()

Hooray, you have just simulated the Earth's atmosphere. Parameters, their meanings and defaults are documented in Parameters. For example, if you want to run the primitive equation dry core (no humidity) simulation in double precision (Float64), at higher resolution (trunc, the triangular spectral truncation), slow down the rotation of the Earth (rotation in $s^{-1}$), and create some netCDF ouput, do

run_speedy(Float64,PrimitiveDryCore,trunc=42,planet=Earth(rotation=1e-5),output=true)

If provided, the number format has to be the first argument, the model (Barotropic, ShallowWater, PrimitiveDryCore, PrimitiveWetCore are available) the second, and all other arguments are keyword arguments.

The run_speedy interface

SpeedyWeather.run_speedy — Function

progn_vars = run_speedy(NF,Model;kwargs...)     or
progn_vars = run_speedy(NF;kwargs...)           or
progn_vars = run_speedy(Model;kwargs...)

Runs SpeedyWeather.jl with number format NF and the model Model and any additional parameters in the keyword arguments kwargs.... Any unspecified parameters will use the default values as defined in Parameters.

The initialize_speedy interface

SpeedyWeather.initialize_speedy — Function

progn_vars, diagn_vars, model_setup = initialize_speedy(NF,Model;kwargs...) or
progn_vars, diagn_vars, model_setup = initialize_speedy(NF,kwargs...)       or
progn_vars, diagn_vars, model_setup = initialize_speedy(Model,kwargs...)

Initialize the model by returning

progn_vars, the initial conditions of the prognostic variables

diagn_vars, the preallocated the diagnotic variables (initialised to zero)

model_setup, the collected pre-calculated structs that don't change throughout integration.

The keyword arguments kwargs are the same as for run_speedy. The model_setup contains fields that hold the parameters, constants, geometry, spectral transform, boundaries and diffusion.