Function and type index

Parameters and constants

SpeedyWeather.Params — Type

Parameter struct that holds all parameters that define the default model setup. With keywords such that default values can be changed at creation.

SpeedyWeather.Constants — Type

Struct holding the parameters needed at runtime in number format NF.

Boundaries

SpeedyWeather.Boundaries — Type

Struct that holds the boundary arrays in grid-point space

geopot_surf     ::Array{Complex{NF},2}  # spectral surface geopotential (orography * gravity) [m^2/s^2]
landsea_mask    ::Array{NF,2}           # land-sea mask, grid-point
albedo          ::Array{NF,2}           # annual mean surface albedo, grid-point

Spectral transform

SpeedyWeather.Geometry — Type

Geometry struct containing parameters and arrays describing the Gaussian grid and the vertical levels. NF is the number format used for the precomputed constants.

SpeedyWeather.fourier — Function

Fourier transform in the zonal direction.

SpeedyWeather.fourier_inverse — Function

Inverse Fourier transform in zonal direction.

SpeedyWeather.legendre — Function

Computes the Legendre transform

SpeedyWeather.legendre_inverse — Function

Computes the inverse Legendre transform.

SpeedyWeather.spectral — Function

Transform a gridded array into spectral space.

SpeedyWeather.gridded — Function

Transform a spectral array into grid-point space.