Function and type index

Parameters and constants

SpeedyWeather.Parameters — Type

Parameter struct that holds all parameters that define the default model setup. With keywords such that default values can be changed at creation.

SpeedyWeather.Constants — Type

Struct holding the parameters needed at runtime in number format NF.

Boundaries and boundary conditions

SpeedyWeather.Boundaries — Type

Struct that holds the boundary arrays in grid-point space geopotsurf ::Array{Complex{NF},2} # spectral surface geopotential (orography * gravity) [m^2/s^2] landseamask ::Array{NF,2} # land-sea mask, grid-point albedo ::Array{NF,2} # annual mean surface albedo, grid-point

Spherical harmonic transform

SpeedyWeather.GeoSpectral — Type

G = Geospectral{NF}(geometry,spectraltrans)

Struct that holds both a Geometry struct geometry and a SpectralTransform struct spectraltrans.

SpeedyWeather.SpectralTransform — Type

SpectralTransform struct that contains all parameters and preallocated arrays for the spectral transform.

SpeedyWeather.spectral — Function

alms = spectral(map)

Forward spectral transform (grid to spectral space) from the gridded field map on a regular Gaussian grid (with Gaussian latitudes) into the spectral coefficients of the Legendre polynomials alms. Based on the size of map this function retrieves the corresponding spectral resolution via triangular truncation and sets up a SpectralTransform struct S to execute spectral(map,S).

alms = spectral(map,S)

Forward spectral transform (grid to spectral space) from the gridded field map on a regular Gaussian grid (with Gaussian latitudes) and the SpectralTransform struct S into the spectral coefficients of the Legendre polynomials alms. This function allocates alms and executes spectral!(alms,map,S).

SpeedyWeather.spectral! — Function

spectral!(alms,map,S)

Forward spectral transform (grid to spectral space) from the gridded field map on a regular Gaussian grid (with Gaussian latitudes). Uses a planned real-valued Fast Fourier Transform in the zonal direction, and a Legendre Transform in the meridional direction exploiting symmetries.Either recomputes the Legendre polynomials Λ for each latitude on one hemisphere or uses precomputed polynomials from S.Λs, depending on S.recompute_legendre. Further uses Legendre weights on Gaussian latitudes for a leakage-free transform.

SpeedyWeather.gridded — Function

map = gridded(alms)

Backward or inverse spectral transform (spectral to grid space) from coefficients alms. Based on the size of alms the corresponding grid space resolution is retrieved based on triangular truncation and a SpectralTransform struct S is allocated to execute gridded(alms,S).

map = gridded(alms,S)

Backward or inverse spectral transform (spectral to grid space) from coefficients alms and the SpectralTransform struct S. Allocates the output map with Gaussian latitudes and executes gridded!(map,alms,S).

SpeedyWeather.gridded! — Function

gridded!(map,alms,S)

Backward or inverse spectral transform (spectral to grid space) from coefficients alms and SpectralTransform struct S into the preallocated output map. Uses a planned inverse Fast Fourier Transform for efficiency in the zonal direction and a Legendre Transform in the meridional direction exploiting symmetries for effciency. Either recomputes the Legendre polynomials to save memory, or uses precomputed polynomials from S depending on S.recompute_legendre.

SpeedyWeather.triangular_truncation — Function

triangular_truncation(trunc::Int,nlon::Int,nlat::Int)

Tests whether the inputs trunc, nlon, nlat satisfy the triangular truncation constraints. trunc is the maximum degree and order of the Legendre polynomials (0-based), nlon is the number of longitudes, nlat the number of latitudes on the spatial grid. The constraints are

- nlon >= 3T+1
- nlat >= (3T+1)/2
- nlon = 2nlat.

nlon, nlat = triangular_truncation(trunc::Int)

Returns the grid size nlon,nlat for a spectral truncation trunc that satisfies the triangular truncation constraints. Returns the smallest pair nlon,nlat that is also easily Fast Fourier-transformable, as determined in roundup_fft.

trunc = triangular_truncation(nlon::Int,nlat::Int)

Returns the largest spectral truncation trunc that satisfies the triangular truncation constraints based on the grid size nlon,nlat, which may or may not be easily Fast Fourier-transformable.

SpeedyWeather.roundup_fft — Function

m = roundup_fft(n::Int)

Returns an integer m >= n with only prime factors 2 and 3 to obtain an efficiently fourier-transformable number of longitudes, m = 2^i * 3^j >= n, with i>=0 but j = 0,1.

SpeedyWeather.spectral_truncation — Function

alms_trunc = spectral_truncation(alms,trunc)

Returns a spectral coefficient matrix alms_trunc that is truncated from alms to the size (trunc+1)^2. alms_trunc only contains those coefficient of alms for which m,l <= trunc, and m>l are zero anyway. If trunc is larger than the implicit truncation in alms obtained from its size than spectral_interpolation is automatically called instead, returning alms_interp, a coefficient matrix that is larger than alms with padded zero coefficients.

SpeedyWeather.spectral_truncation! — Function

spectral_truncation!(alms,trunc)

Truncate spectral coefficients alms in-place by setting (a) the upper right triangle to zero and (b) all coefficients for which the degree l is larger than the truncation trunc.

spectral_truncation!(alms)

Truncate spectral coefficients alms in-place by setting the upper right triangle to zero. This is to enforce that all coefficients for which the degree l is larger than order m are zero.

SpeedyWeather.get_legendre_polynomials! — Function

get_legendre_polynomials!(Λ,Λs,ilat,cos_colat,recompute_legendre)

Base on recompute_legendre (true/false) this function either updates the Legendre polynomials Λ for a given latitude ilat, cos_colat by recomputation (recompute_legendre == true), or Λ is changed by copying over the polynomials from precomputed Λs. Recomputation takes usually longer, but precomputation requires a large amount of memory for high resolution.

SpeedyWeather.∇²! — Function

∇²!(∇²alms::AbstractMatrix{Complex{NF}},    # Output: Laplacian of alms
    alms::AbstractMatrix{Complex{NF}},      # spectral coefficients
    R::Real                                 # radius of the Earth
    ) where {NF<:AbstractFloat}             # number format NF

Spherical Laplace operator ∇² applied to the spectral coefficients alms on a sphere of radius R. The spherical Laplace operator is

∇²alms = -l(l+1)/R²*alms

with the degree l of the Legendre polynomial. ∇²! is the in-place version, storing the result directly in ∇²alms.

∇²!(∇²alms::AbstractMatrix{Complex},    # Output: Laplacian of alms
    alms::AbstractMatrix{Complex})      # spectral coefficients

Same as ∇²!(::AbstractMatrix,::AbstractMatrix,R::Real) but assuming a sphere of radius R=1.

SpeedyWeather.∇² — Function

∇²alms = ∇²(alms::AbstractMatrix{Complex},R::Real=1)

Spherical Laplace operator ∇² applied to the spectral coefficients alms on a sphere of radius R. ∇²(alms,R) is the non-in-place version of ∇²! which first allocates the output array ∇²alms before calling ∇²!.

SpeedyWeather.∇⁴! — Function

∇⁴!(∇⁴alms::AbstractMatrix{Complex},    # Output: Laplacian of alms
    alms::AbstractMatrix{Complex},      # spectral coefficients
    R::Real=1)                          # radius of the sphere/Earth

Spherical Bi-Laplace operator ∇⁴ = ∇²(∇²) applied to the spectral coefficients alms on a sphere of radius R. ∇⁴! operates by applying ∇²! twice. ∇⁴! is the in-place version, storing the result directly in the argument ∇⁴alms.

SpeedyWeather.∇⁴ — Function

∇⁴alms = ∇⁴(alms::AbstractMatrix{Complex},R::Real=1)

Spherical Bi-Laplace operator ∇⁴ = ∇²(∇²) applied to the spectral coefficients alms on a sphere of radius R. ∇⁴ operates by applying ∇² twice. ∇⁴ is the non-in-place version of ∇⁴! which first allocates the output array ∇⁴alms before calling ∇⁴!.

SpeedyWeather.∇⁻²! — Function

∇⁻²!(   ∇⁻²alms::AbstractMatrix{Complex},       # Out: inverse Laplace of alms
        alms::AbstractMatrix{Complex})          # In: spectral coefficients alms

Inverse spherical Laplace operator ∇⁻² applied to the spectral coefficients alms on a sphere of radius R=1. ∇⁻²! is the in-place version which directly stores the output in the argument ∇⁻²alms. The integration constant for Legendre polynomial l=m=0 is zero.

∇⁻²!(   ∇⁻²alms::AbstractMatrix{Complex},
        alms::AbstractMatrix{Complex},
        R::Real=1)

Inverse spherical Laplace operator ∇⁻² applied to the spectral coefficients alms on a sphere of radius R. ∇⁻²! is the in-place version which directly stores the output in the argument ∇⁻²alms. The integration constant for Legendre polynomial l=m=0 is zero. The inverse spherical Laplace operator is

∇⁻²alms = alms*R²/(-l(l+1))

with the degree l (0-based) of the Legendre polynomial.

SpeedyWeather.∇⁻² — Function

∇⁻²alms = ∇⁻²(alms::AbstractMatrix{Complex},R::Real=1)

Inverse spherical Laplace operator ∇⁻² applied to the spectral coefficients alms on a sphere of radius R. ∇⁻² is the non-in-place version of ∇⁻²! and therefore first allocates the output array ∇⁻²alms before calling ∇⁻²!. The integration constant for Legendre polynomial l=m=0 is zero.

SpeedyWeather.gradient_latitude! — Function

cosθ_u = gradient_latitude!(cosθ_u::AbstractArray{Complex{NF}},
                            Ψ::AbstractArray{Complex{NF}},
                            R::Real=1
                            ) where {NF<:AbstractFloat}

Meridional gradient in spectral space of spherical harmonic coefficients Ψ on a sphere with radius R. Returns cosθ_u, i.e. the gradient ∂Ψ/∂θ with an additional cosine of latitude θ scaling. This function uses the recursion relation (0-based degree l, order m)

(cosθ u)_lm = -1/R*(-(l-1)*ϵ_lm*Ψ_l-1,m + (l+2)*ϵ_l+1,m*Ψ_l+1,m ).

As u = -1/R*∂Ψ/∂θ, this function can be generally used to compute the gradient in latitude.

Missing docstring.

Missing docstring for SpeedyWeather.gradient_latitude. Check Documenter's build log for details.

SpeedyWeather.gradient_longitude! — Function

cosθ_∂alms_∂ϕ = gradient_longitude!(    cosθ_∂alms_∂ϕ::AbstractMatrix{Complex{NF}},
                                        alms::AbstractMatrix{Complex{NF}};
                                        R::Real=1
                                        ) where {NF<:AbstractFloat}

Zonal gradient in spectral space of spherical harmonic coefficients alms on a sphere with radius R. While the zonal gradient is 1/cosθ∂/∂ϕ in spherical coordinates, this functions omits the 1/cosθ scaling such that the return array is cosθ∂alms/∂ϕ.

SpeedyWeather.gradient_longitude — Function

Gradient in longitude in spectral space. Input: coefficients alms of the spherical harmonics.

SpeedyWeather.divergence_uvω_spectral — Function

Spectral tendency of ∇⋅(uv*ω) from vector uv=(u,v) in grid space and absolute vorticity ω. Step 1 (grid space): Add Coriolis f to the relative vorticity ζ (=vor_grid) to obtain abs vorticity ω. Step 2 (grid space): Multiply u,v with abs vorticity ω. Step 3 (grid space): Unscale with cosθ, cosine of latitude, as the gradients will include a cosθ term. Step 4 (spectral space): convert uω/cosθ, vω/cosθ from grid to spectral space Step 5 (spectral space): Compute gradients ∂/∂ϕ(uω/cosθ) and ∂/∂θ(vω/cosθ) Step 6 (spectral space): Add ∂/∂ϕ(uω/cosθ)+∂/∂θ(vω/cosθ) and return.

SpeedyWeather.ϵlm — Function

ϵ = ϵ(NF,l,m)

Recursion factors ϵ as a function of degree l and order m (0-based) of the spherical harmonics. ϵ(l,m) = sqrt((l^2-m^2)/(4*l^2-1)) and then converted to number format NF.

ϵ = ϵ(l,m)

Recursion factors ϵ as a function of degree l and order m (0-based) of the spherical harmonics. ϵ(l,m) = sqrt((l^2-m^2)/(4*l^2-1)) with default number format Float64.

Geometry

SpeedyWeather.Geometry — Type

Geometry struct containing parameters and arrays describing the Gaussian grid and the vertical levels. NF is the number format used for the precomputed constants.

SpeedyWeather.vertical_coordinates — Function

Vertical sigma coordinates defined by their nlev+1 half levels σ_levels_half. Sigma coordinates are fraction of surface pressure (p/p0) and are sorted from top (stratosphere) to bottom (surface). The first half level is at 0 the last at 1. Evaluate a generalised logistic function with coefficients in P for the distribution of values in between. Default coefficients follow the L31 configuration historically used at ECMWF.

SpeedyWeather.GenLogisticCoefs — Type

Coefficients of the generalised logistic function to describe the vertical coordinate. Default coefficients A,K,C,Q,B,M,ν are fitted to the old L31 configuration at ECMWF. See geometry.jl and function vertical_coordinate for more informaiton.

Following the notation of https://en.wikipedia.org/wiki/Generalisedlogisticfunction (Dec 15 2021).

Change default parameters for more/fewer levels in the stratosphere vs troposphere vs boundary layer.

SpeedyWeather.generalised_logistic — Function

Generalised logistic function based on the coefficients in coefs.

Time stepping

SpeedyWeather.time_stepping! — Function

Calculate a single time step for SpeedyWeather.jl

Missing docstring.

Missing docstring for SpeedyWeather.time_step!. Check Documenter's build log for details.

SpeedyWeather.first_timestep! — Function

Call initialization of semi-implicit scheme and perform initial time step.

SpeedyWeather.leapfrog! — Function

Perform one leapfrog time step with or without Robert+William's filter (see William (2009), Montly Weather Review, Eq. 7-9)

Leapfrog! for 3D arrays that loops over all vertical layers.