Function and type index

Parameters and constants

SpeedyWeather.Parameters — Type

P = Parameters(kwargs...)

A struct to hold all model parameters that may be changed by the user. The struct uses keywords such that default values can be changed at creation. The default values of the keywords define the default model setup.

SpeedyWeather.Constants — Type

Struct holding the parameters needed at runtime in number format NF.

Boundaries and boundary conditions

SpeedyWeather.Boundaries — Type

Struct that holds the boundary arrays in grid-point space geopotsurf ::Array{Complex{NF},2} # spectral surface geopotential (orography * gravity) [m^2/s^2] landseamask ::Array{NF,2} # land-sea mask, grid-point albedo ::Array{NF,2} # annual mean surface albedo, grid-point

Spherical harmonic transform

Missing docstring.

Missing docstring for SpeedyWeather.GeoSpectral. Check Documenter's build log for details.

SpeedyWeather.SpectralTransform — Type

S = SpectralTransform{NF<:AbstractFloat}(...)

SpectralTransform struct that contains all parameters and preallocated arrays for the spectral transform.

SpeedyWeather.spectral — Function

alms = spectral(    map::AbstractMatrix;
                    Grid::Type{<:AbstractGrid}=FullGaussianGrid,
                    kwargs...)

Converts map to grid(map) to execute spectral(map::AbstractGrid;kwargs...).

alms = spectral(    map::AbstractGrid;
                    Grid::Type{<:AbstractGrid}=FullGaussianGrid,
                    kwargs...)

Converts map to Grid(map) to execute spectral(map::AbstractGrid;kwargs...).

alms = spectral(    map::AbstractMatrix;
                    Grid::Type{<:AbstractGrid}=FullGaussianGrid,
                    kwargs...)

Spectral transform (grid to spectral) map to grid(map) to execute spectral(map::AbstractGrid;kwargs...).

SpeedyWeather.spectral! — Function

spectral!(  alms::LowerTriangularMatrix,
            map::AbstractGrid,
            S::SpectralTransform)

Spectral transform (grid to spectral space) from the gridded field map on a grid<:AbstractGrid to a LowerTriangularMatrix of spherical harmonic coefficients alms. Uses FFT in the zonal direction, and a Legendre Transform in the meridional direction exploiting symmetries. The spectral transform is number format-flexible as long as the parametric types of map, alms, S are identical. The spectral transform is grid-flexible as long as the typeof(map)<:AbstractGrid. Uses the precalculated arrays, FFT plans and other constants in the SpectralTransform struct S.

SpeedyWeather.gridded — Function

map = gridded(  alms::AbstractMatrix;
                recompute_legendre::Bool=true,
                grid::Type{<:AbstractGrid}=FullGaussianGrid)

Spectral transform (spectral to grid space) from spherical coefficients alms to a newly allocated gridded field map. Based on the size of alms the grid type grid, the spatial resolution is retrieved based on the truncation defined for grid. SpectralTransform struct S is allocated to execute gridded(alms,S).

map = gridded(  alms::AbstractMatrix,
                S::SpectralTransform)

Spectral transform (spectral to grid space) from spherical coefficients alms to a newly allocated gridded field map with precalculated properties based on the SpectralTransform struct S. alms is converted to a LowerTriangularMatrix to execute the in-place gridded!.

SpeedyWeather.gridded! — Function

gridded!(   map::AbstractGrid,
            alms::LowerTriangularMatrix,
            S::SpectralTransform)

Spectral transform (spectral to grid) of the spherical harmonic coefficients alms to a gridded field map. The spectral transform is number format-flexible as long as the parametric types of map, alms, S are identical. The spectral transform is grid-flexible as long as the typeof(map)<:AbstractGrid. Uses the precalculated arrays, FFT plans and other constants in the SpectralTransform struct S.

gridded!(   diagn::DiagnosticVariables{NF}, # all diagnostic variables
            progn::PrognosticVariables{NF}, # all prognostic variables
            M::BarotropicModel,             # everything that's constant
            lf::Int=1                       # leapfrog index
            ) where NF

Propagate the spectral state of the prognostic variables progn to the diagnostic variables in diagn for the barotropic vorticity model. Updates grid vorticity, spectral stream function and spectral and grid velocities u,v.

gridded!(   diagn::DiagnosticVariables{NF}, # all diagnostic variables
            progn::PrognosticVariables{NF}, # all prognostic variables
            M::ShallowWaterModel,           # everything that's constant
            lf::Int=1                       # leapfrog index
            ) where NF

Propagate the spectral state of the prognostic variables progn to the diagnostic variables in diagn for the shallow water model. Updates grid vorticity, grid divergence, grid interface displacement (pres_grid) and the velocities U,V (scaled by cos(lat)).

Missing docstring.

Missing docstring for SpeedyWeather.triangular_truncation. Check Documenter's build log for details.

SpeedyWeather.roundup_fft — Function

m = roundup_fft(n::Int;
                small_primes::Vector{Int}=[2,3,5])

Returns an integer m >= n with only small prime factors 2, 3, 5 (default, others can be specified with the keyword argument small_primes) to obtain an efficiently fourier-transformable number of longitudes, m = 2^i * 3^j * 5^k >= n, with i,j,k >=0.

SpeedyWeather.spectral_truncation — Function

alms_trunc = spectral_truncation(alms,trunc)

Returns a spectral coefficient matrix alms_trunc that is truncated from alms to the size (trunc+1)^2. alms_trunc only contains those coefficient of alms for which m,l <= trunc, and l>=m are zero anyway. If trunc is larger than the implicit truncation in alms obtained from its size than spectral_interpolation is automatically called instead, returning alms_interp, a coefficient matrix that is larger than alms with padded zero coefficients. Also works with higher dimensional arrays, but truncation is only applied to the first two dimensions.

SpeedyWeather.spectral_truncation! — Function

spectral_truncation!(alms::AbstractMatrix,ltrunc::Integer,mtrunc::Integer)

Truncate spectral coefficients alms in-place by setting (a) the upper right triangle to zero and (b) all coefficients for which the degree l is larger than the truncation ltrunc or order m larger than the truncaction mtrunc.

spectral_truncation!(alms,trunc)

Truncate spectral coefficients alms in-place by setting (a) the upper right triangle to zero and (b) all coefficients for which the degree l is larger than the truncation trunc.

spectral_truncation!(alms)

Truncate spectral coefficients alms in-place by setting the upper right triangle to zero. This is to enforce that all coefficients for which the degree l is larger than order m are zero.

Missing docstring.

Missing docstring for SpeedyWeather.spectral_interpolation!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.get_legendre_polynomials!. Check Documenter's build log for details.

SpeedyWeather.∇²! — Function

∇²!(    ∇⁻²alms::LowerTriangularMatrix,
        alms::LowerTriangularMatrix,
        S::SpectralTransform)

Laplace operator ∇² applied to the spectral coefficients alms in spherical coordinates. The radius R is omitted in the eigenvalues which are precomputed in S. ∇²! is the in-place version which directly stores the output in the first argument ∇²alms. The spherical Laplace operator is generally

∇⁻²alms = alms*(-l(l+1))/R²

with the degree l (0-based) of the Legendre polynomial.

Missing docstring.

Missing docstring for SpeedyWeather.∇². Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.∇⁻²!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.∇⁻². Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.gradient_latitude!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.gradient_latitude. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.gradient_longitude!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.gradient_longitude. Check Documenter's build log for details.

SpeedyWeather.divergence! — Function

divergence!(div::LowerTriangularMatrix,
            u::LowerTriangularMatrix,
            v::LowerTriangularMatrix,
            S::SpectralTransform{NF};
            flipsign::Bool=false,
            add::Bool=false,
            )

Divergence of a vector u,v written into div, div = ∇⋅(u,v). u,v are expected to have a 1/coslat-scaling included, then div is not scaled. flipsign option calculates -∇⋅(u,v) instead. add option calculates div += ∇⋅(u,v) instead. flipsign and add can be combined. This functions only creates the kernel and calls the generic divergence function _divergence! subsequently.

SpeedyWeather.curl! — Function

curl!(  curl::LowerTriangularMatrix,
        u::LowerTriangularMatrix,
        v::LowerTriangularMatrix,
        S::SpectralTransform;
        flipsign::Bool=false,
        add::Bool=false,
        )

Curl of a vector u,v written into curl, curl = ∇×(u,v). u,v are expected to have a 1/coslat-scaling included, then curl is not scaled. flipsign option calculates -∇×(u,v) instead. add option calculates curl += ∇×(u,v) instead. flipsign and add can be combined. This functions only creates the kernel and calls the generic divergence function _divergence! subsequently with flipped u,v -> v,u for the curl.

SpeedyWeather._divergence! — Function

_divergence!(   kernel,
                div::LowerTriangularMatrix,
                u::LowerTriangularMatrix,
                v::LowerTriangularMatrix,
                S::SpectralTransform)

Generic divergence function of vector u,v that writes into the output into div. Generic as it uses the kernel kernel such that curl, div, add or flipsign options are provided through kernel, but otherwise a single function is used.

Missing docstring.

Missing docstring for SpeedyWeather.curl_div!. Check Documenter's build log for details.

SpeedyWeather.UV_from_vordiv! — Function

UV_from_vordiv!(U::LowerTriangularMatrix,
                V::LowerTriangularMatrix,
                vor::LowerTriangularMatrix,
                div::LowerTriangularMatrix,
                S::SpectralTransform)

Get U,V (=(u,v)*coslat) from vorticity ζ and divergence D in spectral space. Two operations are combined into a single linear operation. First, invert the spherical Laplace ∇² operator to get stream function from vorticity and velocity potential from divergence. Then compute zonal and meridional gradients to get U,V.

SpeedyWeather.UV_from_vor! — Function

UV_from_vor!(   U::LowerTriangularMatrix,
                V::LowerTriangularMatrix,
                vor::LowerTriangularMatrix,
                S::SpectralTransform)

Get U,V (=(u,v)*coslat) from vorticity ζ spectral space (divergence D=0) Two operations are combined into a single linear operation. First, invert the spherical Laplace ∇² operator to get stream function from vorticity. Then compute zonal and meridional gradients to get U,V.

SpeedyWeather.ϵlm — Function

ϵ = ϵ(NF,l,m)

Recursion factors ϵ as a function of degree l and order m (0-based) of the spherical harmonics. ϵ(l,m) = sqrt((l^2-m^2)/(4*l^2-1)) and then converted to number format NF.

ϵ = ϵ(l,m)

Recursion factors ϵ as a function of degree l and order m (0-based) of the spherical harmonics. ϵ(l,m) = sqrt((l^2-m^2)/(4*l^2-1)) with default number format Float64.

SpeedyWeather.get_recursion_factors — Function

get_recursion_factors(  ::Type{NF}, # number format NF
                        lmax::Int,  # max degree l of spherical harmonics (0-based here)
                        mmax::Int   # max order m of spherical harmonics
                        ) where {NF<:AbstractFloat}

Returns a matrix of recursion factors ϵ up to degree lmax and order mmax of number format NF.

Dynamics

SpeedyWeather.bernoulli_potential! — Function

bernoulli_potential!(   D::DiagnosticVariables{NF}, # all diagnostic variables   
                        GS::GeoSpectral{NF},        # struct with geometry and spectral transform
                        g::Real                     # gravity
                        ) where {NF<:AbstractFloat}

Computes the Laplace operator ∇² of the Bernoulli potential B in spectral space. First, computes the Bernoulli potential on the grid, then transforms to spectral space and takes the Laplace operator.

bernoulli_potential!(   B::AbstractGrid,    # Output: Bernoulli potential B = 1/2*(u^2+v^2)+g*η
                        u::AbstractGrid,    # zonal velocity
                        v::AbstractGrid,    # meridional velocity
                        η::AbstractGrid,    # interface displacement
                        g::Real,            # gravity
                        G::Geometry)

Computes the Bernoulli potential 1/2(u^2 + v^2) + gη in grid-point space.

SpeedyWeather.volume_flux_divergence! — Function

volume_fluxes!( D::DiagnosticVariables{NF},
                G::Geometry{NF},
                S::SpectralTransform{NF},
                B::Boundaries,
                H₀::Real                    # layer thickness
                ) where {NF<:AbstractFloat}

Computes the (negative) divergence of the volume fluxes uh,vh for the continuity equation, -∇⋅(uh,vh)

SpeedyWeather.vorticity_fluxes! — Function

vorticity_fluxes!(  uω_coslat⁻¹::AbstractGrid{NF},      # Output: u*(ζ+f)/coslat
                    vω_coslat⁻¹::AbstractGrid{NF},      # Output: v*(ζ+f)/coslat
                    U::AbstractGrid{NF},                # Input: u*coslat
                    V::AbstractGrid{NF},                # Input: v*coslat
                    vor::AbstractGrid{NF},              # Input: relative vorticity ζ
                    G::Geometry{NF}                     # struct with precomputed geometry arrays
                    ) where {NF<:AbstractFloat}         # number format NF

Compute the vorticity fluxes (u,v)*(ζ+f)/coslat in grid-point space from U,V and vorticity ζ.

SpeedyWeather.vorticity_flux_curl! — Function

vorticity_flux_curl!(   D::DiagnosticVariablesLayer,
                        S::SpectralTransform,
                        )

Compute the curl of the vorticity fluxes ∇×(uω,vω) and store in divergence tendency. Requires vorticity_fluxes! to have been calculated already.

SpeedyWeather.vorticity_flux_divergence! — Function

vorticity_flux_divergence!( D::DiagnosticVariables{NF}, # all diagnostic variables   
                            G::GeoSpectral{NF}          # struct with geometry and spectral transform
                            ) where {NF<:AbstractFloat}

Compute the vorticity advection as the (negative) divergence of the vorticity fluxes -∇⋅(uv(ζ+f)). First, compute the uv(ζ+f), then transform to spectral space and take the divergence and flip the sign.

Geometry

SpeedyWeather.Geometry — Type

Geometry{NF<:AbstractFloat}

Geometry struct containing parameters and arrays describing an iso-latitude grid <:AbstractGrid and the vertical levels. NF is the number format used for the precomputed constants.

SpeedyWeather.vertical_coordinates — Function

σ_levels_half = vertical_coordinates(P::Parameters)

Vertical sigma coordinates defined by their nlev+1 half levels σ_levels_half. Sigma coordinates are fraction of surface pressure (p/p0) and are sorted from top (stratosphere) to bottom (surface). The first half level is at 0 the last at 1. Evaluate a generalised logistic function with coefficients in P for the distribution of values in between. Default coefficients follow the L31 configuration historically used at ECMWF.

SpeedyWeather.GenLogisticCoefs — Type

Coefficients of the generalised logistic function to describe the vertical coordinate. Default coefficients A,K,C,Q,B,M,ν are fitted to the old L31 configuration at ECMWF. See geometry.jl and function vertical_coordinate for more informaiton.

Following the notation of https://en.wikipedia.org/wiki/Generalisedlogisticfunction (Dec 15 2021).

Change default parameters for more/fewer levels in the stratosphere vs troposphere vs boundary layer.

SpeedyWeather.generalised_logistic — Function

Generalised logistic function based on the coefficients in coefs.

Time stepping

SpeedyWeather.time_stepping! — Function

time_stepping!( progn::PrognosticVariables,     # all prognostic variables
                diagn::DiagnosticVariables,     # all pre-allocated diagnostic variables
                M::ModelSetup                   # all precalculated structs
                )

Main time loop that that initialises output and feedback, loops over all time steps and calls the output and feedback functions.

SpeedyWeather.timestep! — Function

timestep!(  progn::PrognosticVariables{NF}, # all prognostic variables
            diagn::DiagnosticVariables{NF}, # all pre-allocated diagnostic variables
            dt::Real,                       # time step (mostly =2Δt, but for init steps =Δt,Δt/2)
            M::PrimitiveEquationModel,      # everything that's constant at runtime
            lf1::Int=2,                     # leapfrog index 1 (dis/enables Robert+William's filter)
            lf2::Int=2                      # leapfrog index 2 (time step used for tendencies)
            ) where {NF<:AbstractFloat}

Calculate a single time step for the primitive equation model of SpeedyWeather.jl

timestep!(  progn::PrognosticVariables{NF}, # all prognostic variables
            diagn::DiagnosticVariables{NF}, # all pre-allocated diagnostic variables
            dt::Real,                       # time step (mostly =2Δt, but for init steps =Δt,Δt/2)
            M::ShallowWaterModel,           # everything that's constant at runtime
            lf1::Int=2,                     # leapfrog index 1 (dis/enables Robert+William's filter)
            lf2::Int=2                      # leapfrog index 2 (time step used for tendencies)
            ) where {NF<:AbstractFloat}

Calculate a single time step for the shallow water model of SpeedyWeather.jl

timestep!(  progn::PrognosticVariables,     # all prognostic variables
            diagn::DiagnosticVariables,     # all pre-allocated diagnostic variables
            dt::Real,                       # time step (mostly =2Δt, but for init steps =Δt,Δt/2)
            lf1::Int=2,                     # leapfrog index 1 (dis/enables Robert+William's filter)
            lf2::Int=2,                     # leapfrog index 2 (time step used for tendencies)
            M::BarotropicModel,             # everything that's constant at runtime
            )

Calculate a single time step for the barotropic vorticity equation model of SpeedyWeather.jl

SpeedyWeather.first_timesteps! — Function

first_timesteps!(   progn::PrognosticVariables, # all prognostic variables
                    diagn::DiagnosticVariables, # all pre-allocated diagnostic variables
                    M::ModelSetup,              # everything that is constant at runtime
                    feedback::Feedback          # feedback struct
                    )

Performs the first two initial time steps (Euler forward, unfiltered leapfrog) to populate the prognostic variables with two time steps (t=0,Δt) that can then be used in the normal leap frogging.

SpeedyWeather.leapfrog! — Function

leapfrog!(  A_old::LowerTriangularMatrix{Complex{NF}},      # prognostic variable at t
            A_new::LowerTriangularMatrix{Complex{NF}},      # prognostic variable at t+dt
            tendency::LowerTriangularMatrix{Complex{NF}},   # tendency (dynamics+physics) of A
            dt::Real,                                       # time step (=2Δt, but for init steps =Δt,Δt/2)
            lf::Int=2,                                      # leapfrog index to dis/enable William's filter
            C::Constants{NF},                               # struct with constants used at runtime
            ) where {NF<:AbstractFloat}                     # number format NF

Performs one leapfrog time step with (lf=2) or without (lf=1) Robert+William's filter (see William (2009), Montly Weather Review, Eq. 7-9).