Function and type index

Parameters and constants

Missing docstring.

Missing docstring for SpeedyWeather.Parameters. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.Constants. Check Documenter's build log for details.

Boundaries and boundary conditions

Missing docstring.

Missing docstring for SpeedyWeather.Boundaries. Check Documenter's build log for details.

Spherical harmonic transform

Missing docstring.

Missing docstring for SpeedyWeather.GeoSpectral. Check Documenter's build log for details.

SpeedyWeather.SpeedyTransforms.SpectralTransform — Type

S = SpectralTransform{NF<:AbstractFloat}(...)

SpectralTransform struct that contains all parameters and preallocated arrays for the spectral transform.

SpeedyWeather.SpeedyTransforms.spectral — Function

alms = spectral(    map::AbstractMatrix;
                    Grid::Type{<:AbstractGrid}=DEFAULT_GRID,
                    kwargs...)

Converts map to grid(map) to execute spectral(map::AbstractGrid;kwargs...).

alms = spectral(    map::AbstractGrid;
                    Grid::Type{<:AbstractGrid}=DEFAULT_GRID,
                    kwargs...)

Converts map to Grid(map) to execute spectral(map::AbstractGrid;kwargs...).

alms = spectral(    map::AbstractMatrix;
                    Grid::Type{<:AbstractGrid}=DEFAULT_GRID,
                    kwargs...)

Spectral transform (grid to spectral) map to grid(map) to execute spectral(map::AbstractGrid;kwargs...).

SpeedyWeather.SpeedyTransforms.spectral! — Function

spectral!(  alms::LowerTriangularMatrix,
            map::AbstractGrid,
            S::SpectralTransform)

Spectral transform (grid to spectral space) from the gridded field map on a grid<:AbstractGrid to a LowerTriangularMatrix of spherical harmonic coefficients alms. Uses FFT in the zonal direction, and a Legendre Transform in the meridional direction exploiting symmetries. The spectral transform is number format-flexible as long as the parametric types of map, alms, S are identical. The spectral transform is grid-flexible as long as the typeof(map)<:AbstractGrid. Uses the precalculated arrays, FFT plans and other constants in the SpectralTransform struct S.

SpeedyWeather.SpeedyTransforms.gridded — Function

map = gridded(  alms::AbstractMatrix;
                recompute_legendre::Bool=true,
                grid::Type{<:AbstractGrid}=DEFAULT_GRID)

Spectral transform (spectral to grid space) from spherical coefficients alms to a newly allocated gridded field map. Based on the size of alms the grid type grid, the spatial resolution is retrieved based on the truncation defined for grid. SpectralTransform struct S is allocated to execute gridded(alms,S).

map = gridded(  alms::AbstractMatrix,
                S::SpectralTransform)

Spectral transform (spectral to grid space) from spherical coefficients alms to a newly allocated gridded field map with precalculated properties based on the SpectralTransform struct S. alms is converted to a LowerTriangularMatrix to execute the in-place gridded!.

SpeedyWeather.SpeedyTransforms.gridded! — Function

gridded!(   map::AbstractGrid,
            alms::LowerTriangularMatrix,
            S::SpectralTransform)

Spectral transform (spectral to grid) of the spherical harmonic coefficients alms to a gridded field map. The spectral transform is number format-flexible as long as the parametric types of map, alms, S are identical. The spectral transform is grid-flexible as long as the typeof(map)<:AbstractGrid. Uses the precalculated arrays, FFT plans and other constants in the SpectralTransform struct S.

gridded!(
    diagn::DiagnosticVariables,
    progn::PrognosticVariables,
    lf::Int64,
    model::SpeedyWeather.ModelSetup
)

Propagate the spectral state of progn to diagn using time step/leapfrog index lf. Function barrier that calls gridded! for the respective model.

gridded!(
    diagn::SpeedyWeather.DiagnosticVariablesLayer,
    progn::SpeedyWeather.PrognosticLayerTimesteps,
    lf::Int64,
    model::Barotropic
)

Propagate the spectral state of the prognostic variables progn to the diagnostic variables in diagn for the barotropic vorticity model. Updates grid vorticity, spectral stream function and spectral and grid velocities u,v.

gridded!(
    diagn::SpeedyWeather.DiagnosticVariablesLayer,
    progn::SpeedyWeather.PrognosticLayerTimesteps,
    lf::Int64,
    model::ShallowWater
)

Propagate the spectral state of the prognostic variables progn to the diagnostic variables in diagn for the shallow water model. Updates grid vorticity, grid divergence, grid interface displacement (pres_grid) and the velocities u,v.

gridded!(
    diagn::SpeedyWeather.DiagnosticVariablesLayer,
    progn::SpeedyWeather.PrognosticLayerTimesteps,
    lf::Int64,
    model::PrimitiveEquation
)

Propagate the spectral state of the prognostic variables progn to the diagnostic variables in diagn for primitive equation models. Updates grid vorticity, grid divergence, grid temperature, pressure (pres_grid) and the velocities u,v.

Missing docstring.

Missing docstring for SpeedyWeather.triangular_truncation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.roundup_fft. Check Documenter's build log for details.

SpeedyWeather.SpeedyTransforms.spectral_truncation — Function

alms_trunc = spectral_truncation(alms,trunc)

Returns a spectral coefficient matrix alms_trunc that is truncated from alms to the size (trunc+1)². alms_trunc only contains those coefficient of alms for which m,l ≤ trunc, and l ≥ m are zero anyway. If trunc is larger than the implicit truncation in alms obtained from its size than spectral_interpolation is automatically called instead, returning alms_interp, a coefficient matrix that is larger than alms with padded zero coefficients.

SpeedyWeather.SpeedyTransforms.spectral_truncation! — Function

spectral_truncation!(alms::AbstractMatrix,ltrunc::Integer,mtrunc::Integer)

Truncate spectral coefficients alms in-place by setting (a) the upper right triangle to zero and (b) all coefficients for which the degree l is larger than the truncation ltrunc or order m larger than the truncaction mtrunc.

spectral_truncation!(alms::LowerTriangularMatrix,ltrunc::Integer,mtrunc::Integer)

Truncate spectral coefficients alms in-place by setting all coefficients for which the degree l is larger than the truncation ltrunc or order m larger than the truncaction mtrunc. Similar to spectral_truncation!(::AbstractMatrix, ...) but skips the upper triangle which is zero by design for LowerTriangularMatrix.

spectral_truncation!(alms,trunc)

Truncate spectral coefficients alms in-place by setting (a) the upper right triangle to zero and (b) all coefficients for which the degree l is larger than the truncation trunc.

spectral_truncation!(alms)

Truncate spectral coefficients alms in-place by setting the upper right triangle to zero. This is to enforce that all coefficients for which the degree l is larger than order m are zero.

Missing docstring.

Missing docstring for SpeedyWeather.spectral_interpolation!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.get_legendre_polynomials!. Check Documenter's build log for details.

SpeedyWeather.SpeedyTransforms.∇²! — Function

∇²!(    ∇²alms::LowerTriangularMatrix,
        alms::LowerTriangularMatrix,
        S::SpectralTransform;
        add::Bool=false,
        flipsign::Bool=false,
        inverse::Bool=false)

Laplace operator ∇² applied to the spectral coefficients alms in spherical coordinates. The radius R is omitted in the eigenvalues which are precomputed in S. ∇²! is the in-place version which directly stores the output in the first argument ∇²alms.

Keyword arguments

add=true adds the ∇²(alms) to the output

flipsign=true computes -∇²(alms) instead

inverse=true computes ∇⁻²(alms) instead

Default is add=false, flipsign=false, inverse=false. These options can be combined.

Missing docstring.

Missing docstring for SpeedyWeather.∇². Check Documenter's build log for details.

SpeedyWeather.SpeedyTransforms.∇⁻²! — Function

∇⁻²!(   ∇⁻²alms::LowerTriangularMatrix,
        alms::LowerTriangularMatrix,
        S::SpectralTransform;
        add::Bool=false,
        flipsign::Bool=false)

Calls ∇²!(∇⁻²alms, alms, S; add, flipsign, inverse=true).

Missing docstring.

Missing docstring for SpeedyWeather.∇⁻². Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.gradient_latitude!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.gradient_latitude. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.gradient_longitude!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.gradient_longitude. Check Documenter's build log for details.

SpeedyWeather.SpeedyTransforms.divergence! — Function

divergence!(div::LowerTriangularMatrix,
            u::LowerTriangularMatrix,
            v::LowerTriangularMatrix,
            S::SpectralTransform{NF};
            flipsign::Bool=false,
            add::Bool=false,
            )

Divergence of a vector u,v written into div, div = ∇⋅(u,v). u,v are expected to have a 1/coslat-scaling included, then div is not scaled. flipsign option calculates -∇⋅(u,v) instead. add option calculates div += ∇⋅(u,v) instead. flipsign and add can be combined. This functions only creates the kernel and calls the generic divergence function _divergence! subsequently.

SpeedyWeather.SpeedyTransforms.curl! — Function

curl!(  curl::LowerTriangularMatrix,
        u::LowerTriangularMatrix,
        v::LowerTriangularMatrix,
        S::SpectralTransform;
        flipsign::Bool=false,
        add::Bool=false,
        )

Curl of a vector u,v written into curl, curl = ∇×(u,v). u,v are expected to have a 1/coslat-scaling included, then curl is not scaled. flipsign option calculates -∇×(u,v) instead. add option calculates curl += ∇×(u,v) instead. flipsign and add can be combined. This functions only creates the kernel and calls the generic divergence function _divergence! subsequently with flipped u,v -> v,u for the curl.

Missing docstring.

Missing docstring for SpeedyWeather._divergence!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.curl_div!. Check Documenter's build log for details.

SpeedyWeather.SpeedyTransforms.UV_from_vordiv! — Function

UV_from_vordiv!(U::LowerTriangularMatrix,
                V::LowerTriangularMatrix,
                vor::LowerTriangularMatrix,
                div::LowerTriangularMatrix,
                S::SpectralTransform)

Get U,V (=(u,v)*coslat) from vorticity ζ and divergence D in spectral space. Two operations are combined into a single linear operation. First, invert the spherical Laplace ∇² operator to get stream function from vorticity and velocity potential from divergence. Then compute zonal and meridional gradients to get U,V.

SpeedyWeather.SpeedyTransforms.UV_from_vor! — Function

UV_from_vor!(   U::LowerTriangularMatrix,
                V::LowerTriangularMatrix,
                vor::LowerTriangularMatrix,
                S::SpectralTransform)

Get U,V (=(u,v)*coslat) from vorticity ζ spectral space (divergence D=0) Two operations are combined into a single linear operation. First, invert the spherical Laplace ∇² operator to get stream function from vorticity. Then compute zonal and meridional gradients to get U,V.

Missing docstring.

Missing docstring for SpeedyWeather.ϵlm. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.get_recursion_factors. Check Documenter's build log for details.

Dynamics

SpeedyWeather.bernoulli_potential! — Function

bernoulli_potential!(
    diagn::SpeedyWeather.DiagnosticVariablesLayer{NF, Grid} where Grid<:SpeedyWeather.RingGrids.AbstractGrid{NF},
    S::SpectralTransform
) -> LowerTriangularMatrix{Complex{NF}} where NF<:AbstractFloat

Computes the Laplace operator ∇² of the Bernoulli potential B in spectral space.

computes the kinetic energy KE = ½(u²+v²) on the grid

transforms KE to spectral space

adds geopotential for the Bernoulli potential in spectral space

takes the Laplace operator.

This version is used for both ShallowWater and PrimitiveEquation, only the geopotential calculation in geopotential! differs.

SpeedyWeather.volume_flux_divergence! — Function

volume_flux_divergence!(
    diagn::SpeedyWeather.DiagnosticVariablesLayer,
    surface::SpeedyWeather.SurfaceVariables,
    orog::SpeedyWeather.AbstractOrography,
    constants::DynamicsConstants,
    G::Geometry,
    S::SpectralTransform
)

Computes the (negative) divergence of the volume fluxes uh,vh for the continuity equation, -∇⋅(uh,vh).

Missing docstring.

Missing docstring for SpeedyWeather.vorticity_fluxes!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.vorticity_flux_curl!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.vorticity_flux_divergence!. Check Documenter's build log for details.

Geometry

SpeedyWeather.Geometry — Type

Construct Geometry struct containing parameters and arrays describing an iso-latitude grid <:AbstractGrid and the vertical levels. Pass on SpectralGrid to calculate the following fields

spectral_grid::SpectralGrid: SpectralGrid that defines spectral and grid resolution

Grid::Type{<:SpeedyWeather.RingGrids.AbstractGrid}: grid of the dynamical core Default: spectral_grid.Grid

nlat_half::Int64: resolution parameter nlathalf of Grid, # of latitudes on one hemisphere (incl Equator) Default: spectralgrid.nlat_half

nlon_max::Int64: maximum number of longitudes (at/around Equator) Default: getnlonmax(Grid, nlat_half)

nlon::Int64: =nlonmax, same (used for compatibility), TODO: still needed? Default: nlonmax

nlat::Int64: number of latitude rings Default: getnlat(Grid, nlathalf)

nlev::Int64: number of vertical levels Default: spectral_grid.nlev

npoints::Int64: total number of grid points Default: spectral_grid.npoints

radius::AbstractFloat: Planet's radius [m] Default: spectral_grid.radius

latd::Vector{Float64}: array of latitudes in degrees (90˚...-90˚) Default: getlatd(Grid, nlathalf)

lond::Vector{Float64}: array of longitudes in degrees (0...360˚), empty for non-full grids Default: getlond(Grid, nlathalf)

londs::Vector{NF} where NF<:AbstractFloat: longitude (-180˚...180˚) for each grid point in ring order Default: (getlatdlonds(Grid, nlathalf))[2]

latds::Vector{NF} where NF<:AbstractFloat: latitude (-90˚...˚90) for each grid point in ring order Default: (getlatdlonds(Grid, nlathalf))[1]

sinlat::Vector{NF} where NF<:AbstractFloat: sin of latitudes Default: sind.(latd)

coslat::Vector{NF} where NF<:AbstractFloat: cos of latitudes Default: cosd.(latd)

coslat⁻¹::Vector{NF} where NF<:AbstractFloat: = 1/cos(lat) Default: 1 ./ coslat

coslat²::Vector{NF} where NF<:AbstractFloat: = cos²(lat) Default: coslat .^ 2

coslat⁻²::Vector{NF} where NF<:AbstractFloat: # = 1/cos²(lat) Default: 1 ./ coslat²

σ_levels_half::Vector{NF} where NF<:AbstractFloat: σ at half levels, σk+1/2 Default: spectralgrid.verticalcoordinates.σhalf

σ_levels_full::Vector{NF} where NF<:AbstractFloat: σ at full levels, σₖ Default: 0.5 * (σlevelshalf[2:end] + σlevelshalf[1:end - 1])

σ_levels_thick::Vector{NF} where NF<:AbstractFloat: σ level thicknesses, σₖ₊₁ - σₖ Default: σlevelshalf[2:end] - σlevelshalf[1:end - 1]

ln_σ_levels_full::Vector{NF} where NF<:AbstractFloat: log of σ at full levels, include surface (σ=1) as last element Default: log.(vcat(σlevelsfull, 1))

Missing docstring.

Missing docstring for SpeedyWeather.vertical_coordinates. Check Documenter's build log for details.

SpeedyWeather.GenLogisticCoefs — Type

Coefficients of the generalised logistic function to describe the vertical coordinate. Default coefficients A,K,C,Q,B,M,ν are fitted to the old L31 configuration at ECMWF.

Following the notation of https://en.wikipedia.org/wiki/Generalisedlogisticfunction (Dec 15 2021).

Change default parameters for more/fewer levels in the stratosphere vs troposphere vs boundary layer.

SpeedyWeather.generalised_logistic — Function

Generalised logistic function based on the coefficients in coefs.

Time stepping

SpeedyWeather.time_stepping! — Function

time_stepping!(
    progn::PrognosticVariables,
    diagn::DiagnosticVariables,
    model::SpeedyWeather.ModelSetup
) -> PrognosticVariables

Main time loop that that initializes output and feedback, loops over all time steps and calls the output and feedback functions.

SpeedyWeather.timestep! — Function

timestep!(
    progn::PrognosticVariables,
    diagn::DiagnosticVariables,
    time::Dates.DateTime,
    dt::Real,
    i::Integer,
    model::Barotropic
)
timestep!(
    progn::PrognosticVariables,
    diagn::DiagnosticVariables,
    time::Dates.DateTime,
    dt::Real,
    i::Integer,
    model::Barotropic,
    lf1::Int64
)
timestep!(
    progn::PrognosticVariables,
    diagn::DiagnosticVariables,
    time::Dates.DateTime,
    dt::Real,
    i::Integer,
    model::Barotropic,
    lf1::Int64,
    lf2::Int64
)

Calculate a single time step for the model <: Barotropic.

timestep!(
    progn::PrognosticVariables{NF<:AbstractFloat},
    diagn::DiagnosticVariables{NF<:AbstractFloat, Grid} where Grid<:SpeedyWeather.RingGrids.AbstractGrid{NF<:AbstractFloat},
    time::Dates.DateTime,
    dt::Real,
    i::Integer,
    model::ShallowWater
) -> SpeedyWeather.RingGrids.AbstractGrid{NF} where NF<:AbstractFloat
timestep!(
    progn::PrognosticVariables{NF<:AbstractFloat},
    diagn::DiagnosticVariables{NF<:AbstractFloat, Grid} where Grid<:SpeedyWeather.RingGrids.AbstractGrid{NF<:AbstractFloat},
    time::Dates.DateTime,
    dt::Real,
    i::Integer,
    model::ShallowWater,
    lf1::Int64
) -> SpeedyWeather.RingGrids.AbstractGrid{NF} where NF<:AbstractFloat
timestep!(
    progn::PrognosticVariables{NF<:AbstractFloat},
    diagn::DiagnosticVariables{NF<:AbstractFloat, Grid} where Grid<:SpeedyWeather.RingGrids.AbstractGrid{NF<:AbstractFloat},
    time::Dates.DateTime,
    dt::Real,
    i::Integer,
    model::ShallowWater,
    lf1::Int64,
    lf2::Int64
) -> SpeedyWeather.RingGrids.AbstractGrid{NF} where NF<:AbstractFloat

Calculate a single time step for the model <: ShallowWater.

timestep!(
    progn::PrognosticVariables{NF<:AbstractFloat},
    diagn::DiagnosticVariables{NF<:AbstractFloat, Grid} where Grid<:SpeedyWeather.RingGrids.AbstractGrid{NF<:AbstractFloat},
    time::Dates.DateTime,
    dt::Real,
    i::Integer,
    model::PrimitiveEquation
) -> Any
timestep!(
    progn::PrognosticVariables{NF<:AbstractFloat},
    diagn::DiagnosticVariables{NF<:AbstractFloat, Grid} where Grid<:SpeedyWeather.RingGrids.AbstractGrid{NF<:AbstractFloat},
    time::Dates.DateTime,
    dt::Real,
    i::Integer,
    model::PrimitiveEquation,
    lf1::Int64
) -> Any
timestep!(
    progn::PrognosticVariables{NF<:AbstractFloat},
    diagn::DiagnosticVariables{NF<:AbstractFloat, Grid} where Grid<:SpeedyWeather.RingGrids.AbstractGrid{NF<:AbstractFloat},
    time::Dates.DateTime,
    dt::Real,
    i::Integer,
    model::PrimitiveEquation,
    lf1::Int64,
    lf2::Int64
) -> Any

Calculate a single time step for the model<:PrimitiveEquation

SpeedyWeather.first_timesteps! — Function

first_timesteps!(
    progn::PrognosticVariables,
    diagn::DiagnosticVariables,
    clock::SpeedyWeather.Clock,
    model::SpeedyWeather.ModelSetup,
    output::SpeedyWeather.AbstractOutputWriter
) -> typeof(time)

Performs the first two initial time steps (Euler forward, unfiltered leapfrog) to populate the prognostic variables with two time steps (t=0,Δt) that can then be used in the normal leap frogging.

SpeedyWeather.leapfrog! — Function

leapfrog!(
    A_old::LowerTriangularMatrix{Complex{NF<:AbstractFloat}},
    A_new::LowerTriangularMatrix{Complex{NF<:AbstractFloat}},
    tendency::LowerTriangularMatrix{Complex{NF<:AbstractFloat}},
    dt::Real,
    lf::Int64,
    L::Leapfrog{NF<:AbstractFloat}
)

Performs one leapfrog time step with (lf=2) or without (lf=1) Robert+William's filter (see William (2009), Montly Weather Review, Eq. 7-9).

Longwave radiation

Missing docstring.

Missing docstring for SpeedyWeather.radset!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.radlw_down!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.compute_bbe!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.radlw_up!. Check Documenter's build log for details.

Shortwave radiation

Missing docstring.

Missing docstring for SpeedyWeather.shortwave_radiation!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.solar!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.sol_oz!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.cloud!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpeedyWeather.radsw!. Check Documenter's build log for details.