SpeedyTransforms

SpeedyTransforms is a submodule that has been developed for SpeedyWeather.jl which is technically independent (SpeedyWeather.jl however imports it) and can also be used without running simulations. It is just not put into its own respective repository.

Example transforms

julia> using SpeedyWeather
julia> alms = zeros(ComplexF64,3,3)    # spectral coefficients
julia> alms[2,2] = 1                   # only l=1,m=1 harmonic
julia> map = gridded(alms)             # convert to grid space
8×4 Matrix{Float64}:
 -0.324541  -0.600363  -0.600363  -0.324541
 -0.134429  -0.248678  -0.248678  -0.134429
  0.134429   0.248678   0.248678   0.134429
  0.324541   0.600363   0.600363   0.324541
  0.324541   0.600363   0.600363   0.324541
  0.134429   0.248678   0.248678   0.134429
 -0.134429  -0.248678  -0.248678  -0.134429
 -0.324541  -0.600363  -0.600363  -0.324541
 
julia> spectral(map)                   # back to spectral space
3×3 Matrix{ComplexF64}:
 0.0+0.0im  0.0+0.0im          0.0+0.0im
 0.0+0.0im  1.0+3.60727e-17im  0.0+0.0im
 0.0+0.0im  0.0+0.0im          0.0+0.0im

and we have successfully reobtained the $l=m=1$ spherical harmonic.

Functions and type index

SpeedyWeather.SpeedyTransforms.SpectralTransform — Type

S = SpectralTransform{NF<:AbstractFloat}(...)

SpectralTransform struct that contains all parameters and preallocated arrays for the spectral transform.

SpeedyWeather.SpeedyTransforms.SpectralTransform — Method

S = SpectralTransform(  alms::AbstractMatrix{Complex{NF}};
                        recompute_legendre::Bool=true,
                        Grid::Type{<:AbstractGrid}=DEFAULT_GRID)

Generator function for a SpectralTransform struct based on the size of the spectral coefficients alms and the grid Grid. Recomputes the Legendre polynomials by default.

SpeedyWeather.SpeedyTransforms.SpectralTransform — Method

S = SpectralTransform(NF,Grid,trunc)

Generator function for a SpectralTransform struct. With NF the number format, Grid the grid type <:AbstractGrid and spectral truncation trunc this function sets up necessary constants for the spetral transform. Also plans the Fourier transforms, retrieves the colatitudes, and preallocates the Legendre polynomials (if recompute_legendre == false) and quadrature weights.

SpeedyWeather.SpeedyTransforms.SpectralTransform — Method

S = SpectralTransform(  map::AbstractGrid;
                        recompute_legendre::Bool=true)

Generator function for a SpectralTransform struct based on the size and grid type of gridded field map. Recomputes the Legendre polynomials by default.

SpeedyWeather.SpeedyTransforms.UV_from_vor! — Method

UV_from_vor!(   U::LowerTriangularMatrix,
                V::LowerTriangularMatrix,
                vor::LowerTriangularMatrix,
                S::SpectralTransform)

Get U,V (=(u,v)*coslat) from vorticity ζ spectral space (divergence D=0) Two operations are combined into a single linear operation. First, invert the spherical Laplace ∇² operator to get stream function from vorticity. Then compute zonal and meridional gradients to get U,V.

SpeedyWeather.SpeedyTransforms.UV_from_vordiv! — Method

UV_from_vordiv!(U::LowerTriangularMatrix,
                V::LowerTriangularMatrix,
                vor::LowerTriangularMatrix,
                div::LowerTriangularMatrix,
                S::SpectralTransform)

Get U,V (=(u,v)*coslat) from vorticity ζ and divergence D in spectral space. Two operations are combined into a single linear operation. First, invert the spherical Laplace ∇² operator to get stream function from vorticity and velocity potential from divergence. Then compute zonal and meridional gradients to get U,V.

SpeedyWeather.SpeedyTransforms._divergence! — Method

_divergence!(   kernel,
                div::LowerTriangularMatrix,
                u::LowerTriangularMatrix,
                v::LowerTriangularMatrix,
                S::SpectralTransform)

Generic divergence function of vector u,v that writes into the output into div. Generic as it uses the kernel kernel such that curl, div, add or flipsign options are provided through kernel, but otherwise a single function is used.

SpeedyWeather.SpeedyTransforms.curl! — Method

curl!(  curl::LowerTriangularMatrix,
        u::LowerTriangularMatrix,
        v::LowerTriangularMatrix,
        S::SpectralTransform;
        flipsign::Bool=false,
        add::Bool=false,
        )

Curl of a vector u,v written into curl, curl = ∇×(u,v). u,v are expected to have a 1/coslat-scaling included, then curl is not scaled. flipsign option calculates -∇×(u,v) instead. add option calculates curl += ∇×(u,v) instead. flipsign and add can be combined. This functions only creates the kernel and calls the generic divergence function _divergence! subsequently with flipped u,v -> v,u for the curl.

SpeedyWeather.SpeedyTransforms.divergence! — Method

divergence!(div::LowerTriangularMatrix,
            u::LowerTriangularMatrix,
            v::LowerTriangularMatrix,
            S::SpectralTransform{NF};
            flipsign::Bool=false,
            add::Bool=false,
            )

Divergence of a vector u,v written into div, div = ∇⋅(u,v). u,v are expected to have a 1/coslat-scaling included, then div is not scaled. flipsign option calculates -∇⋅(u,v) instead. add option calculates div += ∇⋅(u,v) instead. flipsign and add can be combined. This functions only creates the kernel and calls the generic divergence function _divergence! subsequently.

SpeedyWeather.SpeedyTransforms.get_recursion_factors — Method

get_recursion_factors(  ::Type{NF}, # number format NF
                        lmax::Int,  # max degree l of spherical harmonics (0-based here)
                        mmax::Int   # max order m of spherical harmonics
                        ) where {NF<:AbstractFloat}

Returns a matrix of recursion factors ϵ up to degree lmax and order mmax of number format NF.

SpeedyWeather.SpeedyTransforms.gridded! — Method

gridded!(   map::AbstractGrid,
            alms::LowerTriangularMatrix,
            S::SpectralTransform)

Spectral transform (spectral to grid) of the spherical harmonic coefficients alms to a gridded field map. The spectral transform is number format-flexible as long as the parametric types of map, alms, S are identical. The spectral transform is grid-flexible as long as the typeof(map)<:AbstractGrid. Uses the precalculated arrays, FFT plans and other constants in the SpectralTransform struct S.

SpeedyWeather.SpeedyTransforms.gridded — Method

map = gridded(  alms::AbstractMatrix;
                recompute_legendre::Bool=true,
                grid::Type{<:AbstractGrid}=DEFAULT_GRID)

Spectral transform (spectral to grid space) from spherical coefficients alms to a newly allocated gridded field map. Based on the size of alms the grid type grid, the spatial resolution is retrieved based on the truncation defined for grid. SpectralTransform struct S is allocated to execute gridded(alms,S).

SpeedyWeather.SpeedyTransforms.gridded — Method

map = gridded(  alms::AbstractMatrix,
                S::SpectralTransform)

Spectral transform (spectral to grid space) from spherical coefficients alms to a newly allocated gridded field map with precalculated properties based on the SpectralTransform struct S. alms is converted to a LowerTriangularMatrix to execute the in-place gridded!.

SpeedyWeather.SpeedyTransforms.is_power_2 — Method

true/false = is_power_2(i::Integer)

Checks whether an integer i is a power of 2, i.e. i = 2^k, with k = 0,1,2,3,....

SpeedyWeather.SpeedyTransforms.roundup_fft — Method

m = roundup_fft(n::Int;
                small_primes::Vector{Int}=[2,3,5])

Returns an integer m >= n with only small prime factors 2, 3 (default, others can be specified with the keyword argument small_primes) to obtain an efficiently fourier-transformable number of longitudes, m = 2^i * 3^j * 5^k >= n, with i,j,k >=0.

SpeedyWeather.SpeedyTransforms.spectral! — Method

spectral!(  alms::LowerTriangularMatrix,
            map::AbstractGrid,
            S::SpectralTransform)

Spectral transform (grid to spectral space) from the gridded field map on a grid<:AbstractGrid to a LowerTriangularMatrix of spherical harmonic coefficients alms. Uses FFT in the zonal direction, and a Legendre Transform in the meridional direction exploiting symmetries. The spectral transform is number format-flexible as long as the parametric types of map, alms, S are identical. The spectral transform is grid-flexible as long as the typeof(map)<:AbstractGrid. Uses the precalculated arrays, FFT plans and other constants in the SpectralTransform struct S.

SpeedyWeather.SpeedyTransforms.spectral — Method

alms = spectral(    map::AbstractMatrix;
                    Grid::Type{<:AbstractGrid}=DEFAULT_GRID,
                    kwargs...)

Converts map to grid(map) to execute spectral(map::AbstractGrid;kwargs...).

SpeedyWeather.SpeedyTransforms.spectral — Method

alms = spectral(    map::AbstractMatrix;
                    Grid::Type{<:AbstractGrid}=DEFAULT_GRID,
                    kwargs...)

Spectral transform (grid to spectral) map to grid(map) to execute spectral(map::AbstractGrid;kwargs...).

SpeedyWeather.SpeedyTransforms.spectral — Method

alms = spectral(    map::AbstractGrid;
                    Grid::Type{<:AbstractGrid}=DEFAULT_GRID,
                    kwargs...)

Converts map to Grid(map) to execute spectral(map::AbstractGrid;kwargs...).

SpeedyWeather.SpeedyTransforms.spectral_interpolation — Method

alms_interp = spectral_interpolation(   ::Type{NF},
                                        alms::LowerTriangularMatrix,
                                        ltrunc::Integer,
                                        mtrunc::Integer
                                        ) where NF

Returns a spectral coefficient matrix alms_interp that is alms padded with zeros to interpolate in spectral space. If trunc is smaller or equal to the implicit truncation in alms obtained from its size than spectral_truncation is automatically called instead, returning alms_trunc, a coefficient matrix that is smaller than alms, implicitly setting higher degrees and orders to zero.

SpeedyWeather.SpeedyTransforms.spectral_smoothing! — Method

spectral_smoothing!(A::LowerTriangularMatrix,c;power=1)

Smooth the spectral field A following A = (1-(1-c)∇²ⁿ) with power n of a normalised Laplacian so that the highest degree lmax is dampened by multiplication with c. Anti-diffusion for c>1.

SpeedyWeather.SpeedyTransforms.spectral_smoothing — Method

A_smooth = spectral_smoothing(A::LowerTriangularMatrix,c;power=1)

Smooth the spectral field A following A_smooth = (1-c*∇²ⁿ)A with power n of a normalised Laplacian so that the highest degree lmax is dampened by multiplication with c. Anti-diffusion for c<0.

SpeedyWeather.SpeedyTransforms.spectral_transform_for_full_grid — Method

S2 = spectral_transform_for_full_grid(S::SpectralTransform)

Create a spectral transform struct S2 similar to the input S, but for the corresponding full grid of the grid in S. The FFT is replanned and lon_offsets are set to 1 (i.e. no rotation). Solid angles for the Legendre transform are recomputed, but all other arrays fields for S, S2 point to the same place in memory, e.g. the Legendre polynomials aren't recomputed or stored twice.

SpeedyWeather.SpeedyTransforms.spectral_truncation! — Method

spectral_truncation!(alms,trunc)

Truncate spectral coefficients alms in-place by setting (a) the upper right triangle to zero and (b) all coefficients for which the degree l is larger than the truncation trunc.

SpeedyWeather.SpeedyTransforms.spectral_truncation! — Method

spectral_truncation!(alms)

Truncate spectral coefficients alms in-place by setting the upper right triangle to zero. This is to enforce that all coefficients for which the degree l is larger than order m are zero.

SpeedyWeather.SpeedyTransforms.spectral_truncation! — Method

spectral_truncation!(alms::AbstractMatrix,ltrunc::Integer,mtrunc::Integer)

Truncate spectral coefficients alms in-place by setting (a) the upper right triangle to zero and (b) all coefficients for which the degree l is larger than the truncation ltrunc or order m larger than the truncaction mtrunc.

SpeedyWeather.SpeedyTransforms.spectral_truncation! — Method

spectral_truncation!(alms::LowerTriangularMatrix,ltrunc::Integer,mtrunc::Integer)

Truncate spectral coefficients alms in-place by setting all coefficients for which the degree l is larger than the truncation ltrunc or order m larger than the truncaction mtrunc. Similar to spectral_truncation!(::AbstractMatrix, ...) but skips the upper triangle which is zero by design for LowerTriangularMatrix.

SpeedyWeather.SpeedyTransforms.spectral_truncation — Method

alms_trunc = spectral_truncation(alms,trunc)

Returns a spectral coefficient matrix alms_trunc that is truncated from alms to the size (trunc+1)². alms_trunc only contains those coefficient of alms for which m,l ≤ trunc, and l ≥ m are zero anyway. If trunc is larger than the implicit truncation in alms obtained from its size than spectral_interpolation is automatically called instead, returning alms_interp, a coefficient matrix that is larger than alms with padded zero coefficients.

SpeedyWeather.SpeedyTransforms.ϵlm — Method

ϵ = ϵ(l,m)

Recursion factors ϵ as a function of degree l and order m (0-based) of the spherical harmonics. ϵ(l,m) = sqrt((l^2-m^2)/(4*l^2-1)) with default number format Float64.

SpeedyWeather.SpeedyTransforms.ϵlm — Method

ϵ = ϵ(NF,l,m)

Recursion factors ϵ as a function of degree l and order m (0-based) of the spherical harmonics. ϵ(l,m) = sqrt((l^2-m^2)/(4*l^2-1)) and then converted to number format NF.

SpeedyWeather.SpeedyTransforms.∇²! — Method

∇²!(    ∇²alms::LowerTriangularMatrix,
        alms::LowerTriangularMatrix,
        S::SpectralTransform;
        add::Bool=false,
        flipsign::Bool=false,
        inverse::Bool=false)

Laplace operator ∇² applied to the spectral coefficients alms in spherical coordinates. The radius R is omitted in the eigenvalues which are precomputed in S. ∇²! is the in-place version which directly stores the output in the first argument ∇²alms.

Keyword arguments

add=true adds the ∇²(alms) to the output

flipsign=true computes -∇²(alms) instead

inverse=true computes ∇⁻²(alms) instead

Default is add=false, flipsign=false, inverse=false. These options can be combined.

SpeedyWeather.SpeedyTransforms.∇⁻²! — Method

∇⁻²!(   ∇⁻²alms::LowerTriangularMatrix,
        alms::LowerTriangularMatrix,
        S::SpectralTransform;
        add::Bool=false,
        flipsign::Bool=false)

Calls ∇²!(∇⁻²alms, alms, S; add, flipsign, inverse=true).